# docs/swagger.yaml
basePath: /
definitions:
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  controllers.defaultResponse:
    properties:
      message:
        type: string
    type: object
...
  models.Product:
    properties:
      category_id:
        type: integer
      description:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Wishlist:
    properties:
      id:
        type: integer
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      user_id:
        type: integer
    type: object
host: localhost:8686
info:
  contact: {}
  description: API Server for Products Application
  title: Products API
  version: "1.0"
paths:
...
  /api/product:
    post:
      consumes:
      - application/json
      description: create new product
      operationId: create-product
      parameters:
      - description: new product info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.defaultResponse'
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Product
      tags:
      - products
  /api/product/{id}:
    delete:
      description: delete product by ID
      operationId: delete-product-by-ID
      parameters:
      - description: id of the product
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.defaultResponse'
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Product By ID
      tags:
      - products
    get:
      consumes:
      - application/json
      description: get product by id
      operationId: get-product-by-id
      parameters:
      - description: id of the product
        in: path
        name: id
        required: true
        type: integer
      - description: product info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.defaultResponse'
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Product By ID
      tags:
      - products
  /api/user/{id}:
    put:
      consumes:
      - application/json
      description: update existed user
      operationId: update-user
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: integer
      - description: user update info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.defaultResponse'
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - users
  /api/users:
    get:
      description: get list of all users
      operationId: get-all-users
      parameters:
      - description: fill if you need search
        in: query
        name: q
        type: string
      - description: fill if you need search
        in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: create new user
      operationId: create-user
      parameters:
      - description: new route info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.defaultResponse'
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create User
      tags:
      - users
  /api/users/{id}:
    get:
      consumes:
      - application/json
      description: get user by id
      operationId: get-user-by-id
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: integer
      - description: user info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.defaultResponse'
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get User By ID
      tags:
      - users
  /api/wishlist:
    post:
      consumes:
      - application/json
      description: create new wishlist
      operationId: create-wishlist
      parameters:
      - description: new wishlist info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Wishlist'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.defaultResponse'
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Wishlist
      tags:
      - wishlists
  /api/wishlist/{id}:
    delete:
      description: remove product from wishlist
      operationId: delete-wishlist-by-ID
      parameters:
      - description: id of the product
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.defaultResponse'
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove From Wishlist
      tags:
      - wishlists
    get:
      consumes:
      - application/json
      description: get wishlist by id
      operationId: get-wishlist-by-id
      parameters:
      - description: id of the wishlist
        in: path
        name: id
        required: true
        type: integer
      - description: wishlist info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Wishlist'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.defaultResponse'
        "400":
          description: Bad Request
          schema:
            type: "404"
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Wishlist By ID
      tags:
      - wishlists
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"